Move loadInstanceProperties to LoadLocalProperties, use in SaveLocalPropertiesIT
